+++ src/Plugin.php
@@ -87,7 +87,7 @@ class Plugin extends \craft\base\Plugin
     /**
      * @inheritdoc
      */
-    public $schemaVersion = '2.3.1';
+    public $schemaVersion = '2.4.0';
 
     /**
      * @var SendRequestJob[] The request jobs that should be queued up at the end of the web request.