+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - Webhooks now requires Craft 3.7 or later.
+- Improved the performance of clearing out completed webhook request logs. ([#32](https://github.com/craftcms/webhooks/issues/32))
 
 ### Fixed
 - Fixed an error that could occur when calling `craft\webhooks\Plugin::getRequestData()` if an invalid request ID was passed.