+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed an error that could occur when calling `craft\webhooks\Plugin::getRequestData()` if an invalid request ID was passed.
 - Fixed an exception that could occur when retrying webhook requests, if the queue driver didn’t support delayed jobs.
 - Fixed an exception that could occur when sending webhook requests, if Guzzle wasn’t able to connect to the server.
+- Fixed an error during install, if the database tables already existed. ([#46](https://github.com/craftcms/webhooks/issues/46))
 
 ## 2.3.3 - 2021-04-01